added report info

+++ src/evaluation/eval.rs
@@ -9,8 +9,9 @@ pub fn material_diff(board: &Board) -> EvalScore {
     let stm = board.stm;
     let mut res = 0;
     for (&pc, &weight) in Piece::LIST.iter().zip(scores.iter()) {
-        let diff = board.piece_bb(pc, stm).popcount() - board.piece_bb(pc, stm.flip()).popcount();
-        res += i32::from(diff) * weight;
+        let diff = i32::from(board.piece_bb(pc, stm).popcount())
+            - i32::from(board.piece_bb(pc, stm.flip()).popcount());
+        res += diff * weight;
     }
     res
 }
